practica1

+++ practica1/conecta_K_2023.c
@@ -142,7 +142,7 @@ void conecta_K_visualizar_tablero(TABLERO *t, uint8_t pantalla[8][8])
 
 
 //
-int conecta_K_verificar_K_en_linea(TABLERO *t, uint8_t fila, uint8_t columna, uint8_t color){
+uint8_t conecta_K_verificar_K_en_linea(TABLERO *t, uint8_t fila, uint8_t columna, uint8_t color){
 	// en esta funcion es donde se debe verificar que todas las optimizaciones dan el mismo resultado
 	uint8_t resultado_c_c = conecta_K_hay_linea_c_c(t, fila, columna, color);
 	uint8_t resultado_c_arm = conecta_K_hay_linea_c_arm(t, fila, columna, color);
